requirements.txt
torch==2.4.1
diffusers==0.32.0
transformers
einops==0.8.0
numpy==2.2.6
peft==0.12.0
matplotlib==3.9.0
pillow
